+++ src/test/java/com/gistpetition/api/integration/user/UserServiceTest.java
@@ -1,10 +1,11 @@
-package com.gistpetition.api.user.application;
+package com.gistpetition.api.integration.user;
 
-import com.gistpetition.api.IntegrationTest;
 import com.gistpetition.api.exception.user.DuplicatedUserException;
 import com.gistpetition.api.exception.user.InvalidEmailFormException;
 import com.gistpetition.api.exception.user.NoSuchUserException;
 import com.gistpetition.api.exception.user.NotMatchedPasswordException;
+import com.gistpetition.api.integration.IntegrationTest;
+import com.gistpetition.api.user.application.UserService;
 import com.gistpetition.api.user.domain.User;
 import com.gistpetition.api.user.domain.UserRepository;
 import com.gistpetition.api.user.domain.UserRole;
