+++ src/test/java/com/gistpetition/api/unit/petition/AgreementsTest.java
@@ -1,6 +1,8 @@
-package com.gistpetition.api.petition.domain;
+package com.gistpetition.api.unit.petition;
 
 import com.gistpetition.api.exception.petition.DuplicatedAgreementException;
+import com.gistpetition.api.petition.domain.Agreement;
+import com.gistpetition.api.petition.domain.Agreements;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
