+++ src/test/java/com/gistpetition/api/unit/utils/BcryptEncoderTest.java
@@ -1,4 +1,4 @@
-package com.gistpetition.api.user;
+package com.gistpetition.api.unit.utils;
 
 import com.gistpetition.api.utils.password.BcryptEncoder;
 import org.junit.jupiter.api.Test;
